Fixed an issue where SSM (SuperbSurvivorManager) was referenced BEFORE it was created...
