+++ mods/SuperbSurvivorsContinued/media/lua/client/04_Group/SuperSurvivorGroup.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager"
+
 SuperSurvivorGroup = {}
 SuperSurvivorGroup.__index = SuperSurvivorGroup
