+++ mods/SuperbSurvivorsContinued/media/lua/client/05_Other/SuperSurvivorSpeakConditions.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 LMSConditions = {}
 LMSConditions.MoodleTable = {}
 LMSConditions.LowerMoodleTable = {}
