+++ mods/SuperbSurvivorsContinued/media/lua/client/06_AI/AI-Manager.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 -- IDEA: If zombies near companion while reloading or something particular, run to player if player is further out
 -- Bug: companions are pursing a target without they returning to the player when out of range
