+++ mods/SuperbSurvivorsContinued/media/lua/client/07_UI/SuperSurvivorInfoPanel.lua
@@ -1,3 +1,4 @@
+require "05_Other/SuperSurvivorManager";
 require "07_UI/UIUtils";
 
 local FONT_HGT_SMALL = getTextManager():getFontHeight(UIFont.Small)
