+++ mods/SuperbSurvivorsContinued/media/lua/client/07_UI/SuperSurvivorLoadout.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 --****************************************************
 -- TitleBar
 --****************************************************
