+++ mods/SuperbSurvivorsContinued/media/lua/client/07_UI/SuperSurvivorMyGroupWindow.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 GroupWindow = ISCollapsableWindow:derive("GroupWindow");
 
 local isLocalLoggingEnabled = false;
