+++ mods/SuperbSurvivorsContinued/media/lua/client/07_UI/UIUtils.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 --****************************************************
 -- RGB functions
 --****************************************************
