+++ mods/SuperbSurvivorsContinued/media/lua/client/08_Tasks/DoctorTask.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 DoctorTask = {}
 DoctorTask.__index = DoctorTask
