+++ mods/SuperbSurvivorsContinued/media/lua/client/08_Tasks/FollowTask.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 FollowTask = {}
 FollowTask.__index = FollowTask
