+++ mods/SuperbSurvivorsContinued/media/lua/client/08_Tasks/PursueTask.lua
@@ -1,3 +1,5 @@
+require "05_Other/SuperSurvivorManager";
+
 PursueTask = {}
 PursueTask.__index = PursueTask
